+++ NOTICE
@@ -1,2 +1,2 @@
-sagemaker-mxnet-serving-container
+sagemaker-mxnet-inference-toolkit
 Copyright 2019 Amazon.com, Inc. or its affiliates. All Rights Reserved.
