+++ artifacts/mms-entrypoint.py
@@ -10,12 +10,17 @@
 # distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF
 # ANY KIND, either express or implied. See the License for the specific
 # language governing permissions and limitations under the License.
-from __future__ import absolute_import
 
-import os
+import shlex
+import subprocess
+import sys
 
-RESOURCE_PATH = os.path.abspath(os.path.join(os.path.dirname(__file__), '..', 'resources'))
 
-# EI is currently only supported in the following regions
-# regions were derived from https://aws.amazon.com/machine-learning/elastic-inference/pricing/
-EI_SUPPORTED_REGIONS = ['us-east-1', 'us-east-2', 'us-west-2', 'eu-west-1', 'ap-northeast-1', 'ap-northeast-2']
+if sys.argv[1] == 'serve':
+    from sagemaker_mxnet_serving_container import serving
+    serving.main()
+else:
+    subprocess.check_call(shlex.split(' '.join(sys.argv[1:])))
+
+# prevent docker exit
+subprocess.call(['tail', '-f', '/dev/null'])
