--- test-toolkit/resources/default_handlers/model/model-shapes.json
@@ -1,1 +1,0 @@
-[{"shape": [1, 2], "name": "data"}]
