--- test-toolkit/resources/default_handlers/model/model-symbol.json
@@ -1,41 +1,0 @@
-{
-  "nodes": [
-    {
-      "op": "null", 
-      "name": "data", 
-      "inputs": []
-    }, 
-    {
-      "op": "null", 
-      "name": "fc1_weight", 
-      "attr": {"num_hidden": "1"}, 
-      "inputs": []
-    }, 
-    {
-      "op": "null", 
-      "name": "fc1_bias", 
-      "attr": {"num_hidden": "1"}, 
-      "inputs": []
-    }, 
-    {
-      "op": "FullyConnected", 
-      "name": "fc1", 
-      "attr": {"num_hidden": "1"}, 
-      "inputs": [[0, 0, 0], [1, 0, 0], [2, 0, 0]]
-    }, 
-    {
-      "op": "null", 
-      "name": "lin_reg_label", 
-      "inputs": []
-    }, 
-    {
-      "op": "LinearRegressionOutput", 
-      "name": "lro", 
-      "inputs": [[3, 0, 0], [4, 0, 0]]
-    }
-  ], 
-  "arg_nodes": [0, 1, 2, 4], 
-  "node_row_ptr": [0, 1, 2, 3, 4, 5, 6], 
-  "heads": [[5, 0, 0]], 
-  "attrs": {"mxnet_version": ["int", 1100]}
-}
