--- test-toolkit/resources/dummy_hosting/code/dummy_hosting_module.py
@@ -1,25 +1,0 @@
-#  Copyright 2019 Amazon.com, Inc. or its affiliates. All Rights Reserved.
-#
-#  Licensed under the Apache License, Version 2.0 (the "License").
-#  You may not use this file except in compliance with the License.
-#  A copy of the License is located at
-#
-#      http://www.apache.org/licenses/LICENSE-2.0
-#
-#  or in the "license" file accompanying this file. This file is distributed
-#  on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either
-#  express or implied. See the License for the specific language governing
-#  permissions and limitations under the License.
-
-
-class DummyModel(object):
-    def predict(self, data):
-        return data
-
-
-def model_fn(model_dir):
-    return DummyModel()
-
-
-def transform_fn(model, data, input_content_type, output_content_type):
-    return data, "application/json"
