+++ test/container/1.6.0/Dockerfile.mxnet.cpu
@@ -58,8 +58,8 @@ RUN useradd -m model-server \
  && mkdir -p /home/model-server/tmp \
  && chown -R model-server /home/model-server
 
-COPY docker/artifacts/mms-entrypoint.py /usr/local/bin/dockerd-entrypoint.py
-COPY docker/artifacts/config.properties /home/model-server
+COPY artifacts/mms-entrypoint.py /usr/local/bin/dockerd-entrypoint.py
+COPY artifacts/config.properties /home/model-server
 
 RUN chmod +x /usr/local/bin/dockerd-entrypoint.py
