--- test/resources/gluon_hosting/model/model.json
@@ -1,126 +1,0 @@
-{
-  "nodes": [
-    {
-      "op": "null", 
-      "name": "data", 
-      "inputs": []
-    }, 
-    {
-      "op": "null", 
-      "name": "sequential0_dense0_weight", 
-      "attr": {
-        "__dtype__": "0", 
-        "__lr_mult__": "1.0", 
-        "__shape__": "(128, 0)", 
-        "__wd_mult__": "1.0"
-      }, 
-      "inputs": []
-    }, 
-    {
-      "op": "null", 
-      "name": "sequential0_dense0_bias", 
-      "attr": {
-        "__dtype__": "0", 
-        "__init__": "zeros", 
-        "__lr_mult__": "1.0", 
-        "__shape__": "(128,)", 
-        "__wd_mult__": "1.0"
-      }, 
-      "inputs": []
-    }, 
-    {
-      "op": "FullyConnected", 
-      "name": "sequential0_dense0_fwd", 
-      "attr": {"num_hidden": "128"}, 
-      "inputs": [[0, 0, 0], [1, 0, 0], [2, 0, 0]]
-    }, 
-    {
-      "op": "Activation", 
-      "name": "sequential0_dense0_relu_fwd", 
-      "attr": {"act_type": "relu"}, 
-      "inputs": [[3, 0, 0]]
-    }, 
-    {
-      "op": "null", 
-      "name": "sequential0_dense1_weight", 
-      "attr": {
-        "__dtype__": "0", 
-        "__lr_mult__": "1.0", 
-        "__shape__": "(64, 0)", 
-        "__wd_mult__": "1.0"
-      }, 
-      "inputs": []
-    }, 
-    {
-      "op": "null", 
-      "name": "sequential0_dense1_bias", 
-      "attr": {
-        "__dtype__": "0", 
-        "__init__": "zeros", 
-        "__lr_mult__": "1.0", 
-        "__shape__": "(64,)", 
-        "__wd_mult__": "1.0"
-      }, 
-      "inputs": []
-    }, 
-    {
-      "op": "FullyConnected", 
-      "name": "sequential0_dense1_fwd", 
-      "attr": {"num_hidden": "64"}, 
-      "inputs": [[4, 0, 0], [5, 0, 0], [6, 0, 0]]
-    }, 
-    {
-      "op": "Activation", 
-      "name": "sequential0_dense1_relu_fwd", 
-      "attr": {"act_type": "relu"}, 
-      "inputs": [[7, 0, 0]]
-    }, 
-    {
-      "op": "null", 
-      "name": "sequential0_dense2_weight", 
-      "attr": {
-        "__dtype__": "0", 
-        "__lr_mult__": "1.0", 
-        "__shape__": "(10, 0)", 
-        "__wd_mult__": "1.0"
-      }, 
-      "inputs": []
-    }, 
-    {
-      "op": "null", 
-      "name": "sequential0_dense2_bias", 
-      "attr": {
-        "__dtype__": "0", 
-        "__init__": "zeros", 
-        "__lr_mult__": "1.0", 
-        "__shape__": "(10,)", 
-        "__wd_mult__": "1.0"
-      }, 
-      "inputs": []
-    }, 
-    {
-      "op": "FullyConnected", 
-      "name": "sequential0_dense2_fwd", 
-      "attr": {"num_hidden": "10"}, 
-      "inputs": [[8, 0, 0], [9, 0, 0], [10, 0, 0]]
-    }
-  ], 
-  "arg_nodes": [0, 1, 2, 5, 6, 9, 10], 
-  "node_row_ptr": [
-    0, 
-    1, 
-    2, 
-    3, 
-    4, 
-    5, 
-    6, 
-    7, 
-    8, 
-    9, 
-    10, 
-    11, 
-    12
-  ], 
-  "heads": [[11, 0, 0]], 
-  "attrs": {"mxnet_version": ["int", 1100]}
-}
